Refactor generators module
