+++ src/aind_data_schema_models/_generated/harp_types.py
@@ -1,13 +1,13 @@
 # generated by aind-data-schema-models:
 #   filename:  https://raw.githubusercontent.com/harp-tech/protocol/97ded281bd1d0d7537f90ebf545d74cf8ba8805e/whoami.yml
-#   timestamp: 2024-08-31 13:49:03.654356
+#   timestamp: 2024-08-31 17:01:14.151435
 
 
 from typing import Annotated, Literal, Union
 
 from pydantic import Field
 
-from aind_data_schema_models.generators import _HarpDeviceTypeModel
+from aind_data_schema_models.generators._base import _HarpDeviceTypeModel
 
 
 class Usbhub(_HarpDeviceTypeModel):
