+++ src/aind_data_schema_models/_generated/modalities.py
@@ -1,13 +1,13 @@
 # generated by aind-data-schema-models:
 #   filename:  modalities.csv
-#   timestamp: 2024-08-31 13:34:29.504806
+#   timestamp: 2024-08-31 17:01:13.919924
 
 
 from typing import Annotated, Literal, Union
 
 from pydantic import Field
 
-from aind_data_schema_models.generators import _ModalityModel
+from aind_data_schema_models.generators._base import _ModalityModel
 
 
 class Behavior(_ModalityModel):
