+++ src/aind_data_schema_models/_generated/platforms.py
@@ -1,13 +1,13 @@
 # generated by aind-data-schema-models:
 #   filename:  platforms.csv
-#   timestamp: 2024-08-31 13:34:29.464055
+#   timestamp: 2024-08-31 17:01:13.878546
 
 
 from typing import Annotated, Literal, Union
 
 from pydantic import Field
 
-from aind_data_schema_models.generators import _PlatformModel
+from aind_data_schema_models.generators._base import _PlatformModel
 
 
 class Behavior(_PlatformModel):
