+++ src/aind_data_schema_models/modalities.py
@@ -1,1 +1,1 @@
-from aind_data_schema_models._generated.modalities import _Modality as Modality # noqa: F401
+from aind_data_schema_models._generated.modalities import _Modality as Modality  # noqa: F401
