+++ src/aind_data_schema_models/platforms.py
@@ -1,1 +1,1 @@
-from aind_data_schema_models._generated.platforms import _Platform as Platform # noqa: F401
+from aind_data_schema_models._generated.platforms import _Platform as Platform  # noqa: F401
